+++ @utahdts/utah-design-system/react/components/forms/RadioButton/context/RadioButtonGroupContextProvider.jsx
@@ -56,7 +56,6 @@ export function RadioButtonGroupContextProvider({
         });
       }
     },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     [value]
   );