+++ @utahdts/utah-design-system/react/components/table/TableFilterTextInput.jsx
@@ -42,8 +42,9 @@ export function TableFilterTextInput({
     onChange,
     value,
   });
+  const { state: { tableWrapperId } } = useTableContext();
 
-  useTableFilterRegistration(recordFieldPath, !!exactMatch);
+  useTableFilterRegistration(recordFieldPath, !!exactMatch, defaultValue);
   const { state: { tableId } } = useTableContext();
 
   return (