+++ @utahdts/utah-design-system/react/components/table/util/TableContext.jsx
@@ -33,5 +33,6 @@ export const TableContext = /** @type {typeof createContext<TableContextValue>} 
     tableSortingOnChange: null,
     tableSortingFieldPath: null,
     tableSortingFieldPaths: null,
+    tableWrapperId: 'default-context-table-wrapper-id',
   },
 });