+++ utah-design-system-website/src/@types/jsDocTypes.d.js
@@ -150,6 +150,39 @@
  * }
  */
 
+/**
+ * @typedef DateInputExamplePropsShape {
+ *  @property {string} className
+ *  @property {string} dateFormat
+ *  @property {string} errorMessage
+ *  @property {boolean} hasCalendarPopup
+ *  @property {string} id
+ *  @property {boolean} isClearable
+ *  @property {boolean} isDisabled
+ *  @property {boolean} isRequired
+ *  @property {string} label
+ *  @property {string} name
+ *  @property {string} placeholder
+ *  @property {boolean} showCalendarTodayButton
+ *  @property {string} value
+ * }
+ */
+
+/**
+ * @typedef FileInputExamplePropsShape {
+ *  @property {string} acceptedFileTypes
+ *  @property {string} className
+ *  @property {string} errorMessage
+ *  @property {string} hint
+ *  @property {string} id
+ *  @property {boolean} isDisabled
+ *  @property {boolean} isRequired
+ *  @property {string} label
+ *  @property {boolean} multiple
+ *  @property {string} name
+ * }
+ */
+
 /** @typedef {Record<CSS_VARIABLES_KEYS, string>} CssContextState */
 
 /**