+++ utah-design-system-website/src/react/components/websiteContent/library/components/forms/FileInput/FileInputDocumentation.jsx
@@ -1,10 +1,15 @@
-/* eslint-disable max-len */
 import { Link } from 'react-router-dom';
+import {
+  FileInput, Tab, TabGroup, TabList, TabPanel, TabPanels
+} from '@utahdts/utah-design-system';
 import { pageUrls } from '../../../../../routing/pageUrls';
 import { StaticExample } from '../../../../../staticExamples/StaticExample';
-import { LightBox } from '../../../../../lightbox/LightBox';
-import fileInputMulti from '../../../../../../../static/images/screenshots/components/form-elements/fileInputMulti.jpg';
-import fileInputSingle from '../../../../../../../static/images/screenshots/components/form-elements/fileInputSingle.jpg';
+import { SandboxExample } from '../../../../../sandbox/SandboxExample';
+import { FileInputExampleCodeReact } from './FileInputExampleCodeReact';
+import { FileInputExampleProps } from './FileInputExampleProps';
+import { FileInputExampleRender } from './FileInputExampleRender';
+import { FileInputCssClassesDocumentation } from './FileInputCssClassesDocumentation';
+import { FileInputPropsDocumentation } from './FileInputPropsDocumentation';
 
 export function FileInputDocumentation() {
   return (
@@ -13,14 +18,32 @@ export function FileInputDocumentation() {
       <p className="lead-in">A file input allows a user to upload and/or attach one or more files.</p>
       <hr />
       <h2 id="example">Example</h2>
+      <SandboxExample
+        defaultProps={{
+          acceptedFileTypes: '',
+          className: '',
+          errorMessage: '',
+          hint: '',
+          id: 'file-input-example',
+          isDisabled: false,
+          isRequired: false,
+          label: 'Choose your file',
+          multiple: false,
+          name: '',
+        }}
+        RENDER_EXAMPLE={FileInputExampleRender}
+        PROPS_EXAMPLE={FileInputExampleProps}
+        CODE_EXAMPLE={FileInputExampleCodeReact}
+      />
+
       <StaticExample
         title="File Input Single"
-        renderedExample={<LightBox image={fileInputSingle} alt="File Input Single" className="flex-3up-gap" />}
+        renderedExample={<div style={{ width: '80%' }}><FileInput id="single-upload-example" label="Upload a single file" /></div>}
       />
 
       <StaticExample
         title="File Input Multi-select"
-        renderedExample={<LightBox image={fileInputMulti} alt="File Input Multi-select" className="flex-3up-gap" />}
+        renderedExample={<div style={{ width: '80%' }}><FileInput id="multiple-upload-example" label="Upload one or more files" multiple /></div>}
       />
 
       <h2 className="mb-spacing" id="section-description-guidance">Description and Guidance</h2>
@@ -89,7 +112,10 @@ export function FileInputDocumentation() {
           the <code>spacebar</code> or <code>enter</code> key should prompt the user to select a file.
         </li>
         <li>The file input should receive focus when the user presses the <code>tab</code> key.</li>
-        <li>Do not only accept files via drag and drop. You must provide a focusable element (button or file input) that the user can trigger using only the keyboard.</li>
+        <li>
+          Do not only accept files via drag and drop.
+          You must provide a focusable element (button or file input) that the user can trigger using only the keyboard.
+        </li>
       </ul>
 
       <h4>Screen Readers</h4>
@@ -104,6 +130,24 @@ export function FileInputDocumentation() {
           to the user on a button element that clearly describes what action they are about to take.
         </li>
       </ul>
+
+      <h2 id="section-settings-props">Settings and Props</h2>
+      <div className="documentation-content--small-text">
+        <TabGroup defaultValue="button-props-css">
+          <TabList>
+            <Tab id="button-props-css">CSS</Tab>
+            <Tab id="button-props-react">React</Tab>
+          </TabList>
+          <TabPanels>
+            <TabPanel tabId="button-props-css">
+              <FileInputCssClassesDocumentation />
+            </TabPanel>
+            <TabPanel tabId="button-props-react">
+              <FileInputPropsDocumentation />
+            </TabPanel>
+          </TabPanels>
+        </TabGroup>
+      </div>
     </div>
   );
 }